requirements.txt
fastapi
uvicorn

langchain
langchain-openai
langchain-community
langgraph

openai
requests
python-dotenv

arize-phoenix
arize-phoenix-client
arize-phoenix-evals
arize-phoenix-otel

openinference-instrumentation
openinference-instrumentation-openai

pandas